[ci] Reduce CI build time
For all CRA projects, the build process involves basic linting and webpack building. Webpack building may run some TSC check.
For all docusaurus projects, the build process only involves webpack building.
Currently, the CI is wasting a lot of time building stuff that is known to be good. The only exception is docusaurus building. (Since it's still in alpha, nothing is so sure.)
Therefore, I introduced a `compile` job to replace the build job. It only runs `tsc` for CRA projects but fallbacks to `yarn build` for docusaurus projects. Now we no longer need to pay for the same spent on running useless minimization for CI build.